% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{as_sem}
\alias{as_sem}
\title{Convert output from package dsem to sem}
\usage{
as_sem(object, lag = 0)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{lag}{what lag to extract and visualize}
}
\value{
Convert output to format supplied by \code{\link[sem]{sem}}
}
\description{
Convert dsem to sem output
}
