% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{fixed}
\alias{fixed}
\title{Family for data that are known without error}
\usage{
fixed()
}
\arguments{
\item{link}{Link for first part of delta/hurdle model.}
}
\description{
Allows using \code{family = fixed()} to specify data that have no measurement error
}
