% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hare_lynx}
\alias{hare_lynx}
\title{Lynx-Hare dynamics}
\usage{
data(hare_lynx)
}
\description{
Data used to demonstrate and test nonlinear dynamics
}
\details{
records of pelts for Canada Lynx and their prey snowshoe hare from
Hudson Bay 1900-1920, extracted from Gotelli (2008 Fig. 6.16) and
originating elsewhere (Elton & Nicholson, 1942; MacLulich, 1937)
}
\references{
Gotelli, N. J. (2008). A Primer of Ecology, Fourth Edition by Nicholas J. Gotelli. Sinauer Associates, 2008.

Elton, C., & Nicholson, M. (1942). The Ten-Year Cycle in Numbers of the
Lynx in Canada. Journal of Animal Ecology, 11(2), 215-244. \doi{10.2307/1358}

MacLulich, D. A. (1937). Fluctuations in the Numbers of the Varying Hare
(Lepus Americanus). University of Toronto Press. \doi{10.3138/9781487583064}
}
\keyword{data}
