% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isle_royale}
\alias{isle_royale}
\title{Isle Royale wolf and moose}
\usage{
data(isle_royale)
}
\description{
Data used to demonstrate and test cross-lagged (vector autoregressive) models
}
\keyword{data}
