% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_parameters.R
\name{list_parameters}
\alias{list_parameters}
\title{List fixed and random effects}
\usage{
list_parameters(Obj, verbose = TRUE)
}
\arguments{
\item{Obj}{Compiled TMB object}

\item{verbose}{Boolean, whether to print messages to terminal}
}
\value{
Tagged-list of fixed and random effects, returned invisibly and printed to screen
}
\description{
\code{list_parameters} lists all fixed and random effects
}
