% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{logLik.dsem}
\alias{logLik.dsem}
\title{Marglinal log-likelihood}
\usage{
\method{logLik}{dsem}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{...}{Not used}
}
\value{
object of class \code{logLik} with attributes
  \item{val}{log-likelihood}
  \item{df}{number of parameters}

Returns an object of class logLik. This has attributes
"df" (degrees of freedom) giving the number of (estimated) fixed effects
in the model, abd "val" (value) giving the marginal log-likelihood.
This class then allows \code{AIC} to work as expected.
}
\description{
Extract the (marginal) log-likelihood of a dsem model
}
