% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{lognormal}
\alias{lognormal}
\title{Lognormal distribution}
\usage{
lognormal(link)
}
\arguments{
\item{link}{Link function}
}
\description{
Lognormal distribution
}
\details{
Central tendancy meanlog is the median of log-response
i.e., does not include bias correction term $sigma^2/2$
}
