% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_dfa.R
\name{make_dfa}
\alias{make_dfa}
\title{Make text for dynamic factor analysis}
\usage{
make_dfa(variables, n_factors, factor_names = paste0("F", seq_len(n_factors)))
}
\arguments{
\item{variables}{Character string of variables (i.e., column names of \code{tsdata}).}

\item{n_factors}{Number of factors.}

\item{factor_names}{Optional character-vector of factor names,
which must match NA columns in \code{tsdata}.}
}
\value{
A text string to be passed to \code{\link{dsem}}
}
\description{
Make the text string for a dynamic factor analysis expressed using
arrow-and-lag notation for DSEM.
}
