% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{print.dsem}
\alias{print.dsem}
\title{Print fitted dsem object}
\usage{
\method{print}{dsem}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{dsem}}}

\item{...}{Not used}
}
\value{
No return value, called to provide clean terminal output when calling fitted
object in terminal.
}
\description{
Prints output from fitted dsem model
}
