% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_model.R
\name{read_model}
\alias{read_model}
\title{Make a RAM (Reticular Action Model)}
\usage{
read_model(sem, times, variables, covs = NULL, quiet = FALSE)
}
\arguments{
\item{sem}{Specification for time-series structural equation model structure
including lagged or simultaneous effects.  See Details section in
\code{\link[dsem]{make_dsem_ram}} for more description}

\item{times}{A character vector listing the set of times in order}

\item{variables}{A character vector listing the set of variables}

\item{covs}{A character vector listing variables for which to estimate a standard deviation}

\item{quiet}{Boolean indicating whether to print messages to terminal}
}
\description{
\code{read_model} converts SEM arrow notation to \code{model} describing SEM parameters
}
\details{
See \code{\link{make_dsem_ram}} for details
}
