% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{residuals.dsem}
\alias{residuals.dsem}
\title{Calculate residuals for dsem}
\usage{
\method{residuals}{dsem}(object, type = c("deviance", "response"), ...)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{type}{which type of residuals to compute (only option is \code{"deviance"} or \code{"response"} for now)}

\item{...}{Note used}
}
\description{
Calculate residuals
}
