% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{simulate.dsem}
\alias{simulate.dsem}
\title{Simulate from a fitted \code{dsem} model}
\usage{
\method{simulate}{dsem}(
  object,
  nsim = 1,
  seed = NULL,
  parametric_uncertainty = c("none", "random", "both"),
  ...
)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{nsim}{number of simulated data sets}

\item{seed}{random seed}

\item{parametric_uncertainty}{whether to ignore uncertainty in fixed and
random effects, include estimation uncertainty in random effects,
or include estimation uncertainty in both fixed and random effects}

\item{...}{Not used}
}
\description{
This function conducts a parametric bootstrap, i.e., simulates new data
conditional upon estimated values for fixed and random effects.  The user
can optionally simulate new random effects conditional upon their estimated
covariance, or simulate new fixed and random effects conditional upon their imprecision.
}
\details{
Simulate dsem
}
