% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{summary.dsem}
\alias{summary.dsem}
\title{Summarize dsem}
\usage{
\method{summary}{dsem}(object, ...)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{...}{Note used}
}
\description{
summarize dsem
}
