% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{as_fitted_DAG}
\alias{as_fitted_DAG}
\title{Convert output from package dsem to phylopath}
\usage{
as_fitted_DAG(fit, lag = 0, what = "Estimate", direction = 1)
}
\arguments{
\item{fit}{Output from \code{\link{dsem}}}

\item{lag}{which lag to output}

\item{what}{whether to output estimates \code{what="Estimate"}, standard errors \code{what="Std_Error"}
or p-values \code{what="Std_Error"}}

\item{direction}{whether to include one-sided arrows \code{direction=1}, or both one- and two-sided arrows \code{direction=c(1,2)}}
}
\value{
Convert output to format supplied by \code{\link[phylopath]{est_DAG}}
}
\description{
Convert dsem to phylopath output
}
