% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{dsem}
\alias{dsem}
\title{Fit dynamic structural equation model}
\usage{
dsem(
  sem,
  tsdata,
  family = Map(function(.) fixed(), colnames(tsdata)),
  estimate_delta0 = FALSE,
  estimate_mu = NULL,
  prior_negloglike = NULL,
  control = dsem_control(),
  covs = colnames(tsdata)
)
}
\arguments{
\item{sem}{Specification for time-series structural equation model structure
including lagged or simultaneous effects.  See Details section in
\code{\link[dsem]{make_dsem_ram}} for more description}

\item{tsdata}{time-series data, as outputted using \code{\link[stats]{ts}}, with \code{NA}
for missing values.}

\item{family}{A named list of families, each returning a class \code{family},
including \code{\link[=fixed]{fixed()}}, \code{\link[=gaussian]{gaussian()}}, \code{\link[=binomial]{binomial()}},  \code{\link[=Gamma]{Gamma()}}, \code{\link[=poisson]{poisson()}},
with names that match levels of
\code{colnames(tsdata)} to allow different
families by variable.  Family \code{\link[=fixed]{fixed()}} specifies that states
are known (i.e., measurements for that variable have no error).
Other families allow users to supply a link function including \code{identity},
\code{log}, \code{logit}, or \code{cloglog}.  For example
\code{family = list(y = binomial("logit"), x = fixed())}
would specify logit-linked Bernoulli distribution for variable \code{tsdata$y}
and a fixed (no measurement error) distribution for \code{tsdata$x}.
For many variables, it is convenient to do e.g.,
\code{family = Map(function(.) gaussian(), colnames(tsdata))} rather than writing
them all manually.}

\item{estimate_delta0}{Boolean indicating whether to estimate deviations from equilibrium in initial year
as fixed effects, or alternatively to assume that dynamics start at some stochastic draw away from
the stationary distribution}

\item{estimate_mu}{character-vector listing columns of \code{tsdata} for which to estimate the mean,
which is subtracted off of \code{tsdata} prior to evaluating interactions among parameters.
The default \code{estimate_mu = NULL} estimates the mean for every column with at least one value
that is not \code{NA} (i.e., does \emph{not} estimate the mean for latent variables).
If you want to have no \code{mu} parameters, use \code{estimate_mu = vector()}.}

\item{prior_negloglike}{A user-provided function that takes as input the vector of fixed effects out$obj$par
returns the negative log-prior probability. For example
\code{prior_negloglike = function(obj) -1 * dnorm( obj$par[1], mean=0, sd=0.1, log=TRUE)}
specifies a normal prior probability
for the for the first fixed effect with mean of zero and logsd of 0.1.
NOTE:  this implementation does not work well with \code{tmbstan} and
is highly experimental.  Note that the user must load RTMB using
\code{library(RTMB)} prior to running the model.}

\item{control}{Output from \code{\link{dsem_control}}, used to define user
settings, and see documentation for that function for details.}

\item{covs}{optional: a character vector of one or more elements, with each element giving a string of variable
names, separated by commas. Variances and covariances among all variables in each such string are
added to the model. Warning: covs="x1, x2" and covs=c("x1", "x2") are not equivalent:
covs="x1, x2" specifies the variance of x1, the variance of x2, and their covariance,
while covs=c("x1", "x2") specifies the variance of x1 and the variance of x2 but not their covariance.
These same covariances can be added manually via argument \code{sem}, but using argument \code{covs} might
save time for models with many variables.}
}
\value{
An object (list) of class \code{dsem}. Elements include:
\describe{
\item{obj}{TMB object from \code{\link[TMB]{MakeADFun}}}
\item{ram}{RAM parsed by \code{make_dsem_ram}}
\item{model}{SEM structure parsed by \code{make_dsem_ram} as intermediate description of model linkages}
\item{tmb_inputs}{The list of inputs passed to \code{\link[TMB]{MakeADFun}}}
\item{opt}{The output from \code{\link[stats]{nlminb}}}
\item{sdrep}{The output from \code{\link[TMB]{sdreport}}}
\item{interal}{Objects useful for package function, i.e., all arguments
passed during the call}
\item{run_time}{Total time to run model}
}
}
\description{
Fits a dynamic structural equation model
}
\details{
A DSEM involves (at a minimum):
\describe{
\item{Time series}{a matrix \eqn{\mathbf X} where column \eqn{\mathbf x_c} for variable c is
a time-series;}
\item{Path diagram}{a user-supplied specification for the path coefficients, which
define the precision (inverse covariance) \eqn{\mathbf Q} for a matrix of state-variables
and see \code{\link{make_dsem_ram}} for more details on the math involved.}
}
The model also estimates the time-series mean \eqn{ \mathbf{\mu}_c } for each variable.
The mean and precision matrix therefore define a Gaussian Markov random field for \eqn{\mathbf X}:

\deqn{ \mathrm{vec}(\mathbf X) \sim \mathrm{MVN}( \mathrm{vec}(\mathbf{I_T} \otimes \mathbf{\mu}), \mathbf{Q}^{-1}) }

Users can the specify
a distribution for measurement errors (or assume that variables are measured without error) using
argument \code{family}.  This defines the link-function \eqn{g_c(.)} and distribution \eqn{f_c(.)}
for each time-series \eqn{c}:

\deqn{ y_{t,c} \sim f_c( g_c^{-1}( x_{t,c} ), \theta_c )}

\code{dsem} then estimates all specified coefficients, time-series means \eqn{\mu_c}, and distribution
measurement errors \eqn{\theta_c} via maximizing a log-marginal likelihood, while
also estimating state-variables \eqn{x_{t,c}}.
\code{summary.dsem} then assembles estimates and standard errors in an easy-to-read format.
Standard errors for fixed effects (path coefficients, exogenoux variance parameters, and measurement error parameters)
are estimated from the matrix of second derivatives of the log-marginal likelihod,
and standard errors for random effects (i.e., missing or state-space variables) are estimated
from a generalization of this method (see \code{\link[TMB]{sdreport}} for details).

\strong{Latent variables}

Any column \eqn{\mathbf x_c} of \code{tsdata} that includes only \code{NA} values
represents a latent variable, and all others are called manifest variables.
The identifiability criteria for latent variables
can be complicated.  To explain, we ignore lagged effects (only simultaneous paths)
and classify three types of latent variables:
\describe{
\item{factor latent variables: }{any latent variable \eqn{\mathbf F} that includes paths out from it
to manifest variables, but has no paths from manifest variables into \eqn{\mathbf F} is a
factor variable.  These are identifable by fixing their SD (i.e., at one), and using a
trimmed Cholesky parameterization (i.e., each successive factor includes fewer paths to
manifest variables).  See the DFA vignette for an example.  Factor latent variables
can be used to represent residual covariance while also estimating the source of
that covariance explicitly}
\item{intermediate latent variables: }{Any latent variable \eqn{\mathbf Y} that includes paths in
from some manifest variables \eqn{\mathbf X} and some paths out to manifest variables
\eqn{\mathbf Z} is an intermediate latent variable.  In general, the at least one path
in or out must be fixed a priori (e.g., at one) to identify the scale of the intermediate
LV.  These intermediate latent variables can represent ecological concepts that serve
as intermediate link between different manifest variables}
\item{composite latent variables: }{Any latent variable \eqn{\mathbf C} that includes paths in
from some manifest variables \eqn{\mathbf X} and no paths out to manifest variables
is a composite latent variable.  In general, you must fix all paths to composite variables
a priori, and must also fix the SD a priori (e.g., at zero).  These composite variables
allow DSEM to estimate a response with standard errors that integrates
across multiple manifest variables}
}
As stated, these criteria do not involve paths from one to another latent variable.  These
are also possible, but involve more complicated identifiability criteria.

\strong{When to do (ot not do) model selection}

In general, DSEM can be used for predictive modelling and/or
structural causal modelling.

For predictive modelling, DSEM provides an expressive
interface to specify any number of fixed effects and use these to represent the
covariance among variables and over time.  The predictive error is expected to
decrease when using a parsimonious model, and model selection might be appropriate
using either \code{\link{stepwise_selection}} or some manual rule for dropping
coefficients that are not statistically significant using a likelihood ratio or
Wald test.

However, structural causal modelling (SCM) is necessary for models to be transferable
to new environments (patterns of colinearity), or for counterfactual analysis.
In general, SCM does not involve using parsimony as a basis for model selection.
Instead, SCM structure should be defined based on ecological knowledge, and
models can be further elaborated using tests of directional separation
(see \code{\link{test_dsep}}).
}
\examples{
# Define model
sem = "
  # Link, lag, param_name
  cprofits -> consumption, 0, a1
  cprofits -> consumption, 1, a2
  pwage -> consumption, 0, a3
  gwage -> consumption, 0, a3
  cprofits -> invest, 0, b1
  cprofits -> invest, 1, b2
  capital -> invest, 0, b3
  gnp -> pwage, 0, c2
  gnp -> pwage, 1, c3
  time -> pwage, 0, c1
"

# Load data
data(KleinI, package="AER")
TS = ts(data.frame(KleinI, "time"=time(KleinI) - 1931))
tsdata = TS[,c("time","gnp","pwage","cprofits",'consumption',
               "gwage","invest","capital")]

# Fit model
fit = dsem(
  sem = sem,
  tsdata = tsdata,
  estimate_delta0 = TRUE,
  control = dsem_control(quiet=TRUE)
)
summary( fit )
plot( fit )
plot( fit, edge_label="value" )

}
\references{
\strong{Introducing the package, its features, and comparison with other software
(to cite when using dsem):}

Thorson, J. T., Andrews, A., Essington, T., Large, S. (2024).
Dynamic structural equation models synthesize
ecosystem dynamics constrained by ecological mechanisms.
Methods in Ecology and Evolution. \doi{10.1111/2041-210X.14289}
}
