% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{dsem}
\alias{dsem}
\title{Fit dynamic structural equation model}
\usage{
dsem(
  sem,
  tsdata,
  family = rep("fixed", ncol(tsdata)),
  estimate_delta0 = FALSE,
  quiet = FALSE,
  run_model = TRUE,
  use_REML = TRUE,
  parameters = NULL,
  map = NULL,
  ...
)
}
\arguments{
\item{sem}{structural equation model structure, passed to either \code{\link[sem]{specifyModel}}
or \code{\link[sem]{specifyEquations}} and then parsed to control
the set of path coefficients and variance-covariance parameters}

\item{tsdata}{time-series data, as outputted using \code{\link[stats]{ts}}}

\item{family}{Character-vector listing the distribution used for each column of \code{tsdata}, where
each element must be \code{fixed} or \code{normal}.
\code{family="fixed"} is default behavior and assumes that a given variable is measured exactly.
Other options correspond to different specifications of measurement error.}

\item{estimate_delta0}{Boolean indicating whether to estimate deviations from equilibrium in initial year
as fixed effects, or alternatively to assume that dynamics start at some stochastic draw away from
the stationary distribution}

\item{quiet}{Boolean indicating whether to run model printing messages to terminal or not;}

\item{run_model}{Boolean indicating whether to estimate parameters (the default), or
instead to return the model inputs and compiled TMB object without running;}

\item{use_REML}{Boolean indicating whether to treat non-variance fixed effects as random,
either to motigate bias in estimated variance parameters or improve efficiency for
parameter estimation given correlated fixed and random effects}

\item{parameters}{list of fixed and random effects, e.g., as constructed by \code{dsem} and then modified
by hand (only helpful for advanced users to change starting values or restart at intended values)}

\item{map}{list of fixed and mirrored parameters, constructed by \code{dsem} by default but available
to override this default and then pass to \code{\link[TMB]{MakeADFun}}}

\item{...}{Additional parameters passed to \code{\link{fit_tmb}}}
}
\value{
An object (list) of class `dsem`. Elements include:
\describe{
\item{obj}{TMB object from \code{\link[TMB]{MakeADFun}}}
\item{ram}{RAM parsed by \code{make_ram}}
\item{model}{SEM model parsed from \code{sem} using \code{\link[sem]{specifyModel}} or \code{\link[sem]{specifyEquations}}}
\item{tmb_inputs}{The list of inputs passed to \code{\link[TMB]{MakeADFun}}}
\item{opt}{The output from \code{\link{fit_tmb}}}
}
}
\description{
Fits a dynamic structural equation model
}
\examples{
# Define model
sem = "
  Profits -> Consumption, 0, a2
  Profits -> Consumption, -1, a3
  Priv_wage -> Consumption, 0, a4
  Gov_wage -> Consumption, 0, a4
  Consumption <-> Consumption, 0, v1
  Consumption -> Consumption, -1, ar1
  Consumption -> Consumption, -2, ar2
  Profits -> Investment, 0, b2
  Profits -> Investment, -1, b3
  Capital_stock -> Investment, -1, b4
  Investment <-> Investment, 0, v2
  neg_Gov_wage <-> neg_Gov_wage, 0, v3
  GNP -> Priv_wage, 0, c2
  Taxes -> Priv_wage, 0, c2
  neg_Gov_wage -> Priv_wage, 0, c2
  GNP -> Priv_wage, -1, c3
  Taxes -> Priv_wage, -1, c3
  neg_Gov_wage -> Priv_wage, -1, c3
  Time -> Priv_wage, 0, c4
  Priv_wage <-> Priv_wage, 0, v4
  GNP <-> GNP, 0, v5
  Profits <-> Profits, 0, v6
  Capital_stock <-> Capital_stock, 0, v7
  Taxes <-> Taxes, 0, v8
  Time <-> Time, 0, v9
  Gov_wage <-> Gov_wage, 0, v10
  Gov_expense <-> Gov_expense, 0, v11
"

# Load data
data(KleinI, package="AER")
Data = as.data.frame(KleinI)
Data = cbind( Data, "time" = seq(1,22)-11 )
colnames(Data) = sapply( colnames(Data), FUN=switch,
           "consumption"="Consumption", "invest"="Investment",
           "cprofits"="Profits", "capital"="Capital_stock", "gwage"="Gov_wage",
           "pwage"="Priv_wage", "gexpenditure"="Gov_expense", "taxes"="Taxes",
           "time"="Time", "gnp"="GNP")
Z = ts( cbind(Data, "neg_Gov_wage"=-1*Data[,'Gov_wage']) )

# Fit model
fit = dsem( sem=sem, tsdata=Z )
summary( fit )

# Plot results
library(ggplot2)
library(ggpubr)
library(phylopath)
p1 = plot(as_fitted_DAG(fit), text_size=3, type="width", show.legend=FALSE)
p1$layers[[1]]$mapping$edge_width = 0.5
p2 = plot(as_fitted_DAG(fit, lag=-1), text_size=3, type="width", show.legend=FALSE)
p2$layers[[1]]$mapping$edge_width = 0.25
ggarrange(p1 + scale_x_continuous(expand = c(0.2, 0.0)),
                    p2 + scale_x_continuous(expand = c(0.2, 0.0)),
                    labels = c("Simultaneous effects", "Lag-1 effects"),
                    ncol = 1, nrow = 2)

}
\references{
**Introducing the package, its features, and comparison with other software
(to cite when using dsem):**

Thorson, J. T., Andrews, A., Essington, T., Large, S. (In review).
Dynamic structural equation models synthesize
ecosystem dynamics constrained by ecological mechanisms.
}
