% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/family.R
\name{gaussian_fixed_sd}
\alias{gaussian_fixed_sd}
\title{Gaussian with known standard deviation for measurement errors}
\usage{
gaussian_fixed_sd(link, sd)
}
\arguments{
\item{link}{Link function}

\item{sd}{vector of known standard deviations (repeated for length of time-series, so be careful)}
}
\description{
Allows using \code{family = gaussian_fixed_sd(link, sd)} to specify data that have known measurement error
}
