% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{isle_royale}
\alias{isle_royale}
\title{Isle Royale wolf and moose}
\usage{
data(isle_royale)
}
\description{
Data used to demonstrate and test cross-lagged (vector autoregressive) models
}
\details{
Data extracted from file "Data_wolves_moose_Isle_Royale_June2019.csv" available at
\url{https://www.isleroyalewolf.org} and obtained 2023-06-23.
Reproduced with permission from John Vucetich, and generated by the
Wolves and Moose of Isle Royale project.
}
\references{
Vucetich, JA and Peterson RO. 2012. The population biology of Isle Royale wolves and moose: an overview. \url{https://www.isleroyalewolf.org}
}
\keyword{data}
