% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lake_washington}
\alias{lake_washington}
\title{Lake washington plankton}
\usage{
data(lake_washington)
}
\description{
Data used to demonstrate and test statistical interactions
}
\details{
measurements of Temperature (W_t in Celcius), Cryptomonas (resource, C_t in log-abundance),
Daphnia (consumer, D_t in log-abundance), and Leptodora (predator, L_t in log-abundance)
in Lake Washington from 1962-1994, T=396 .
}
\references{
Hampton, S. E., Scheuerell, M. D., & Schindler, D. E. (2006). Coalescence in the Lake Washington story:
Interaction strengths in a planktonic food web. Limnology and Oceanography, 51(5),
2042-2051. \doi{10.4319/lo.2006.51.5.2042}
}
\keyword{data}
