% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_msv.R
\name{make_msv}
\alias{make_msv}
\title{Make text for multivariate stochastic volatility model}
\usage{
make_msv(
  variables,
  n_factors,
  factor_names = paste0("F", seq_len(n_factors)),
  collapse_text = TRUE
)
}
\arguments{
\item{variables}{Character string of variables (i.e., column names of \code{tsdata}).}

\item{n_factors}{Number of stochastic volatility factors.}

\item{factor_names}{Optional character-vector of factor names,
which must match NA columns in \code{tsdata}.}

\item{collapse_text}{whether to collapse text into long character string}
}
\value{
A text string to be passed to \code{\link{dsem}}
}
\description{
Make the text string for a multivariate stochastic volatility model using
arrow-lag-slope notation for DSEM.
}
