% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paramesium_didinium}
\alias{paramesium_didinium}
\title{Paramesium-Didinium dynamics}
\usage{
data(paramesium_didinium)
}
\description{
Data used to demonstrate and test nonlinear dynamics
}
\details{
records of Paramesium aurelia and Didinium nasutum in a microcosm experiment
at 0.5 Cerophyll concentration measured every 12 hours over 35 days, i.e., T=71
(Veilleux, 1976 Fig. 11a), as previously digitized (Jost & Ellner, 2000 Fig. 1).
}
\references{
Veilleux, B. G. (1979). An Analysis of the Predatory Interaction Between Paramecium
and Didinium. Journal of Animal Ecology, 48(3), 787-803. \doi{10.2307/4195}

Jost, C., & Ellner, S. P. (2000). Testing for predator dependence in predator-prey
dynamics: A non-parametric approach. Proceedings of the Royal Society of London.
Series B: Biological Sciences. \doi{10.1098/rspb.2000.1186}
}
\keyword{data}
