% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdo_departure_bay}
\alias{pdo_departure_bay}
\title{Pacific Decadal Oscillation and Departure Bay temperatures}
\usage{
data(pdo_departure_bay)
}
\description{
Data used to demonstrate and test nonstationary dynamics
}
\details{
records of the Pacific Decadal Oscillation by year (row) and month
(Jan-Dec, columns), downloaded from JISAO \url{http://research.jisao.washington.edu/pdo/PDO.latest.txt}
on Nov. 6, 2018, as analyzed previously by Thorson et al. (2020)
}
\references{
Thorson, J. T., Ciannelli, L., & Litzow, M. A. (2020). Defining indices of
ecosystem variability using biological samples of fish communities:
A generalization of empirical orthogonal functions. Progress in Oceanography,
181, 102244. \doi{10.1016/j.pocean.2019.102244}
}
\keyword{data}
