% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dsem.R
\name{residuals.dsem}
\alias{residuals.dsem}
\title{Calculate residuals}
\usage{
\method{residuals}{dsem}(object, type = c("deviance", "response"), ...)
}
\arguments{
\item{object}{Output from \code{\link{dsem}}}

\item{type}{which type of residuals to compute (only option is \code{"deviance"} or \code{"response"} for now)}

\item{...}{Not used}
}
\value{
A matrix of residuals, with same order and dimensions as argument \code{tsdata}
that was passed to \code{dsem}.
}
\description{
Calculate deviance or response residuals for dsem
}
