% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stepwise_selection.R
\name{stepwise_selection}
\alias{stepwise_selection}
\title{Simulate dsem}
\usage{
stepwise_selection(model_options, model_shared, quiet = FALSE, ...)
}
\arguments{
\item{model_options}{character-vector containing sem elements
that could be included or dropped depending upon their
parsimony}

\item{model_shared}{character-vector containing sem elements
that must be included regardless of parsimony}

\item{quiet}{whether to avoid displaying progress to terminal}

\item{...}{arguments passed to \code{\link{dsem}},
other than \code{sem} e.g., \code{tsdata}, \code{family}
etc.}
}
\value{
An object (list) that includes:
\describe{
\item{model}{the string with the selected SEM model}
\item{record}{a list showing the AIC and whether each \code{model_options} is included or not}
}
}
\description{
Plot from a fitted \code{dsem} model
}
\details{
This function conducts stepwise (i.e., forwards and backwards) model
selection using marginal AIC, while forcing some model elements to be
included and selecting among others.
}
