% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_dsep.R
\name{test_dsep}
\alias{test_dsep}
\title{Test d-separation}
\usage{
test_dsep(object, max_lag = NULL)
}
\arguments{
\item{object}{object from \code{\link{dsem}}}

\item{max_lag}{how many lags to include when defining the set of conditional
independence relationships. If missing, this value is taken from
the maximum lag that's included in the model.}
}
\value{
A p-value representing the weight of evidence that the data arises
from the specified model, where a low p-value indicates
significant evidence for rejecting this hypothesis.
}
\description{
Calculate the p-value for a test of d-separation \strong{(Experimental)}
}
\details{
A user-specified SEM implies a set of conditional independence relationships
among variables, which can be fitted individually, extracting the
slope and associated p-value, and then combining these p-values to define
a model-wide (omnibus) p-value for the hypothesis that a given data set arises
from the specified model.  This test is modified from package:phylopath.  However
it is unclear exactly how to define the set of conditional-independence assumptions
in a model with temporal autocorrelation, and the test was not developed for
uses when data are missing.  At the time of writing, the function is hightly
experimental.
}
\references{
Shipley, B. (2000). A new inferential test
for path models based on directed acyclic graphs. Structural
  Equation Modeling, 7(2), 206-218. \doi{10.1207/S15328007SEM0702_4}
}
